USB: Elevate USB HAL service privilige to root
usb-hal-1-0 service need to read&write the device node
which user and group is root. usb-hal-1-0 does not have
the privilege due its user and group is system. This HAL
is required for the framework to access the states of
the USB port.

Hence elevate usb_hal_1-0 service to root to make sure the
service can access all the device node with enough privilege.

Change-Id: Id96ae212ea2226cf30cf51074715b1d32fe9b044